pyproject.toml
[project]
name = "auditshield-live"
version = "0.1.0"
description = "AuditShield-Live — Medicare Advantage RADV Audit Defense Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "shiny>=0.7.1",
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "plotly>=5.18.0",
    "anthropic>=0.39.0",
    "scikit-learn>=1.4.0",
    "psycopg2-binary>=2.9.9",
    "gspread>=6.0.0",
    "supabase>=2.10.0",
    "google-auth>=2.28.0",
    "xlsxwriter>=3.1.9",
    "python-dotenv>=1.0.0",
    "faker>=22.6.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest-playwright>=0.4.0",
]

# Entry: app:app (shiny run app.py) or api:app (uvicorn api:app) for /request-trial
# Port: 7860 (Shiny) or 8000 (API with trial endpoint)

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
exclude = ["starguard-desktop", "starguard-mobile", "auditshield-live"]

[tool.ruff.lint.per-file-ignores]
"app.py" = ["E402"]  # Deferred imports (Phase 1/2/3 blocks)

[tool.ruff.isort]
known-first-party = ["app", "audit_trail", "cloud_status_badge"]

[tool.mypy]
python_version = "3.11"
mypy_path = "."
explicit_package_bases = true
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
